Added database connection to the logger.
Fixed a bug where the dict created using the old data was added to the database as the dict was created. That meant first adding sensor (0,) then (0,1,), then (0,1,2,) and so on.

Also moved the creation of the tuple (ts,) inside the actual database method. Now db.addData(ts,data) takes ts as a string argument (much easier to handle)